pyproject.toml
[project]
name = "seasonal-shift"
version = "0.1.0"
description = "A tool to rename and move episode files."
authors = [{ name = "Youssef Aswad", email = "youssefaswad@gmail.com" }]
license = "MIT"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["pyyaml (>=6.0.3,<7.0.0)", "typer (>=0.24.1,<0.25.0)", "rich (>=14.3.3,<15.0.0)", "pydantic (>=2.12.5,<3.0.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
  "basedpyright (>=1.38.2,<2.0.0)",
  "black (>=26.3.0,<27.0.0)",
  "isort (>=8.0.1,<9.0.0)",
  "autoflake (>=2.3.3,<3.0.0)",
]

[tool.poetry]
packages = [{ include = "seasonal_shift", from = "src" }]

[project.scripts]
seasonal-shift = "seasonal_shift.main:app"
